Add Warning About Installation Issues
I came across Pytorch Lighting Flash but was unable to install it. The community figured out what the issues are and it is something that others will face too. Until it is fixed I recommend including a link to the issue and workarounds so that other people are not stuck. See https://github.com/PyTorchLightning/lightning-flash/issues/803